added constants for more ciphers
