+++ ssl/cipher_suites.go
@@ -1,0 +1,383 @@
+package ssl
+
+const (
+	// TLS 1.0 - 1.2 cipher suites.
+	TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256     uint16 = 0xcca9
+	TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256       uint16 = 0xcca8
+	TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256_OLD uint16 = 0xcc14
+	TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256_OLD   uint16 = 0xcc13
+	TLS_DHE_RSA_WITH_CHACHA20_POLY1305_SHA256_OLD     uint16 = 0xcc15
+	TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384             uint16 = 0xc030
+	TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384           uint16 = 0xc02c
+	TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384             uint16 = 0xc028
+	TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384           uint16 = 0xc024
+	TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA                uint16 = 0xc014
+	TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA              uint16 = 0xc00a
+	TLS_SRP_SHA_DSS_WITH_AES_256_CBC_SHA              uint16 = 0xc022
+	TLS_SRP_SHA_RSA_WITH_AES_256_CBC_SHA              uint16 = 0xc021
+	TLS_SRP_SHA_WITH_AES_256_CBC_SHA                  uint16 = 0xc020
+	TLS_DH_DSS_WITH_AES_256_GCM_SHA384                uint16 = 0x00a5
+	TLS_DHE_DSS_WITH_AES_256_GCM_SHA384               uint16 = 0x00a3
+	TLS_DH_RSA_WITH_AES_256_GCM_SHA384                uint16 = 0x00a1
+	TLS_DHE_RSA_WITH_AES_256_GCM_SHA384               uint16 = 0x009f
+	TLS_DHE_RSA_WITH_AES_256_CBC_SHA256               uint16 = 0x006b
+	TLS_DHE_DSS_WITH_AES_256_CBC_SHA256               uint16 = 0x006a
+	TLS_DH_RSA_WITH_AES_256_CBC_SHA256                uint16 = 0x0069
+	TLS_DH_DSS_WITH_AES_256_CBC_SHA256                uint16 = 0x0068
+	TLS_DHE_RSA_WITH_AES_256_CBC_SHA                  uint16 = 0x0039
+	TLS_DHE_DSS_WITH_AES_256_CBC_SHA                  uint16 = 0x0038
+	TLS_DH_RSA_WITH_AES_256_CBC_SHA                   uint16 = 0x0037
+	TLS_DH_DSS_WITH_AES_256_CBC_SHA                   uint16 = 0x0036
+	TLS_ECDHE_RSA_WITH_CAMELLIA_256_CBC_SHA384        uint16 = 0xc077
+	TLS_ECDHE_ECDSA_WITH_CAMELLIA_256_CBC_SHA384      uint16 = 0xc073
+	TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA256          uint16 = 0x00c4
+	TLS_DHE_DSS_WITH_CAMELLIA_256_CBC_SHA256          uint16 = 0x00c3
+	TLS_DH_RSA_WITH_CAMELLIA_256_CBC_SHA256           uint16 = 0x00c2
+	TLS_DH_DSS_WITH_CAMELLIA_256_CBC_SHA256           uint16 = 0x00c1
+	TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA             uint16 = 0x0088
+	TLS_DHE_DSS_WITH_CAMELLIA_256_CBC_SHA             uint16 = 0x0087
+	TLS_DH_RSA_WITH_CAMELLIA_256_CBC_SHA              uint16 = 0x0086
+	TLS_DH_DSS_WITH_CAMELLIA_256_CBC_SHA              uint16 = 0x0085
+	TLS_ECDH_anon_WITH_AES_256_CBC_SHA                uint16 = 0xc019
+	TLS_DH_anon_WITH_AES_256_GCM_SHA384               uint16 = 0x00a7
+	TLS_DH_anon_WITH_AES_256_CBC_SHA256               uint16 = 0x006d
+	TLS_DH_anon_WITH_AES_256_CBC_SHA                  uint16 = 0x003a
+	TLS_DH_anon_WITH_CAMELLIA_256_CBC_SHA256          uint16 = 0x00c5
+	TLS_DH_anon_WITH_CAMELLIA_256_CBC_SHA             uint16 = 0x0089
+	TLS_ECDH_RSA_WITH_AES_256_GCM_SHA384              uint16 = 0xc032
+	TLS_ECDH_ECDSA_WITH_AES_256_GCM_SHA384            uint16 = 0xc02e
+	TLS_ECDH_RSA_WITH_AES_256_CBC_SHA384              uint16 = 0xc02a
+	TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA384            uint16 = 0xc026
+	TLS_ECDH_RSA_WITH_AES_256_CBC_SHA                 uint16 = 0xc00f
+	TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA               uint16 = 0xc005
+	TLS_ECDH_RSA_WITH_CAMELLIA_256_CBC_SHA384         uint16 = 0xc079
+	TLS_ECDH_ECDSA_WITH_CAMELLIA_256_CBC_SHA384       uint16 = 0xc075
+	TLS_RSA_WITH_AES_256_GCM_SHA384                   uint16 = 0x009d
+	TLS_RSA_WITH_AES_256_CBC_SHA256                   uint16 = 0x003d
+	TLS_RSA_WITH_AES_256_CBC_SHA                      uint16 = 0x0035
+	TLS_RSA_WITH_CAMELLIA_256_CBC_SHA256              uint16 = 0x00c0
+	TLS_RSA_WITH_CAMELLIA_256_CBC_SHA                 uint16 = 0x0084
+	TLS_RSA_PSK_WITH_AES_256_CBC_SHA                  uint16 = 0x0095
+	TLS_PSK_WITH_AES_256_CBC_SHA                      uint16 = 0x008d
+	TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             uint16 = 0xc02f
+	TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256           uint16 = 0xc02b
+	TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256             uint16 = 0xc027
+	TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256           uint16 = 0xc023
+	TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA                uint16 = 0xc013
+	TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA              uint16 = 0xc009
+	TLS_SRP_SHA_DSS_WITH_AES_128_CBC_SHA              uint16 = 0xc01f
+	TLS_SRP_SHA_RSA_WITH_AES_128_CBC_SHA              uint16 = 0xc01e
+	TLS_SRP_SHA_WITH_AES_128_CBC_SHA                  uint16 = 0xc01d
+	TLS_DH_DSS_WITH_AES_128_GCM_SHA256                uint16 = 0x00a4
+	TLS_DHE_DSS_WITH_AES_128_GCM_SHA256               uint16 = 0x00a2
+	TLS_DH_RSA_WITH_AES_128_GCM_SHA256                uint16 = 0x00a0
+	TLS_DHE_RSA_WITH_AES_128_GCM_SHA256               uint16 = 0x009e
+	TLS_DHE_RSA_WITH_AES_128_CBC_SHA256               uint16 = 0x0067
+	TLS_DHE_DSS_WITH_AES_128_CBC_SHA256               uint16 = 0x0040
+	TLS_DH_RSA_WITH_AES_128_CBC_SHA256                uint16 = 0x003f
+	TLS_DH_DSS_WITH_AES_128_CBC_SHA256                uint16 = 0x003e
+	TLS_DHE_RSA_WITH_AES_128_CBC_SHA                  uint16 = 0x0033
+	TLS_DHE_DSS_WITH_AES_128_CBC_SHA                  uint16 = 0x0032
+	TLS_DH_RSA_WITH_AES_128_CBC_SHA                   uint16 = 0x0031
+	TLS_DH_DSS_WITH_AES_128_CBC_SHA                   uint16 = 0x0030
+	TLS_ECDHE_RSA_WITH_CAMELLIA_128_CBC_SHA256        uint16 = 0xc076
+	TLS_ECDHE_ECDSA_WITH_CAMELLIA_128_CBC_SHA256      uint16 = 0xc072
+	TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA256          uint16 = 0x00be
+	TLS_DHE_DSS_WITH_CAMELLIA_128_CBC_SHA256          uint16 = 0x00bd
+	TLS_DH_RSA_WITH_CAMELLIA_128_CBC_SHA256           uint16 = 0x00bc
+	TLS_DH_DSS_WITH_CAMELLIA_128_CBC_SHA256           uint16 = 0x00bb
+	TLS_DHE_RSA_WITH_SEED_CBC_SHA                     uint16 = 0x009a
+	TLS_DHE_DSS_WITH_SEED_CBC_SHA                     uint16 = 0x0099
+	TLS_DH_RSA_WITH_SEED_CBC_SHA                      uint16 = 0x0098
+	TLS_DH_DSS_WITH_SEED_CBC_SHA                      uint16 = 0x0097
+	TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA             uint16 = 0x0045
+	TLS_DHE_DSS_WITH_CAMELLIA_128_CBC_SHA             uint16 = 0x0044
+	TLS_DH_RSA_WITH_CAMELLIA_128_CBC_SHA              uint16 = 0x0043
+	TLS_DH_DSS_WITH_CAMELLIA_128_CBC_SHA              uint16 = 0x0042
+	TLS_ECDH_anon_WITH_AES_128_CBC_SHA                uint16 = 0xc018
+	TLS_DH_anon_WITH_AES_128_GCM_SHA256               uint16 = 0x00a6
+	TLS_DH_anon_WITH_AES_128_CBC_SHA256               uint16 = 0x006c
+	TLS_DH_anon_WITH_AES_128_CBC_SHA                  uint16 = 0x0034
+	TLS_DH_anon_WITH_CAMELLIA_128_CBC_SHA256          uint16 = 0x00bf
+	TLS_DH_anon_WITH_SEED_CBC_SHA                     uint16 = 0x009b
+	TLS_DH_anon_WITH_CAMELLIA_128_CBC_SHA             uint16 = 0x0046
+	TLS_ECDH_RSA_WITH_AES_128_GCM_SHA256              uint16 = 0xc031
+	TLS_ECDH_ECDSA_WITH_AES_128_GCM_SHA256            uint16 = 0xc02d
+	TLS_ECDH_RSA_WITH_AES_128_CBC_SHA256              uint16 = 0xc029
+	TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA256            uint16 = 0xc025
+	TLS_ECDH_RSA_WITH_AES_128_CBC_SHA                 uint16 = 0xc00e
+	TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA               uint16 = 0xc004
+	TLS_ECDH_RSA_WITH_CAMELLIA_128_CBC_SHA256         uint16 = 0xc078
+	TLS_ECDH_ECDSA_WITH_CAMELLIA_128_CBC_SHA256       uint16 = 0xc074
+	TLS_RSA_WITH_AES_128_GCM_SHA256                   uint16 = 0x009c
+	TLS_RSA_WITH_AES_128_CBC_SHA256                   uint16 = 0x003c
+	TLS_RSA_WITH_AES_128_CBC_SHA                      uint16 = 0x002f
+	TLS_RSA_WITH_CAMELLIA_128_CBC_SHA256              uint16 = 0x00ba
+	TLS_RSA_WITH_SEED_CBC_SHA                         uint16 = 0x0096
+	TLS_RSA_WITH_CAMELLIA_128_CBC_SHA                 uint16 = 0x0041
+	TLS_RSA_WITH_IDEA_CBC_SHA                         uint16 = 0x0007
+	TLS_RSA_PSK_WITH_AES_128_CBC_SHA                  uint16 = 0x0094
+	TLS_PSK_WITH_AES_128_CBC_SHA                      uint16 = 0x008c
+	TLS_ECDHE_RSA_WITH_RC4_128_SHA                    uint16 = 0xc011
+	TLS_ECDHE_ECDSA_WITH_RC4_128_SHA                  uint16 = 0xc007
+	TLS_DHE_DSS_WITH_RC4_128_SHA                      uint16 = 0x0066
+	TLS_ECDH_anon_WITH_RC4_128_SHA                    uint16 = 0xc016
+	TLS_DH_anon_WITH_RC4_128_MD5                      uint16 = 0x0018
+	TLS_ECDH_RSA_WITH_RC4_128_SHA                     uint16 = 0xc00c
+	TLS_ECDH_ECDSA_WITH_RC4_128_SHA                   uint16 = 0xc002
+	TLS_RSA_WITH_RC4_128_SHA                          uint16 = 0x0005
+	TLS_RSA_WITH_RC4_128_MD5                          uint16 = 0x0004
+	TLS_RSA_PSK_WITH_RC4_128_SHA                      uint16 = 0x0092
+	TLS_PSK_WITH_RC4_128_SHA                          uint16 = 0x008a
+	TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA               uint16 = 0xc012
+	TLS_ECDHE_ECDSA_WITH_3DES_EDE_CBC_SHA             uint16 = 0xc008
+	TLS_SRP_SHA_DSS_WITH_3DES_EDE_CBC_SHA             uint16 = 0xc01c
+	TLS_SRP_SHA_RSA_WITH_3DES_EDE_CBC_SHA             uint16 = 0xc01b
+	TLS_SRP_SHA_WITH_3DES_EDE_CBC_SHA                 uint16 = 0xc01a
+	TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA                 uint16 = 0x0016
+	TLS_DHE_DSS_WITH_3DES_EDE_CBC_SHA                 uint16 = 0x0013
+	TLS_DH_RSA_WITH_3DES_EDE_CBC_SHA                  uint16 = 0x0010
+	TLS_DH_DSS_WITH_3DES_EDE_CBC_SHA                  uint16 = 0x000d
+	TLS_ECDH_anon_WITH_3DES_EDE_CBC_SHA               uint16 = 0xc017
+	TLS_DH_anon_WITH_3DES_EDE_CBC_SHA                 uint16 = 0x001b
+	TLS_ECDH_RSA_WITH_3DES_EDE_CBC_SHA                uint16 = 0xc00d
+	TLS_ECDH_ECDSA_WITH_3DES_EDE_CBC_SHA              uint16 = 0xc003
+	TLS_RSA_WITH_3DES_EDE_CBC_SHA                     uint16 = 0x000a
+	TLS_RSA_PSK_WITH_3DES_EDE_CBC_SHA                 uint16 = 0x0093
+	TLS_PSK_WITH_3DES_EDE_CBC_SHA                     uint16 = 0x008b
+	TLS_DHE_DSS_EXPORT1024_WITH_DES_CBC_SHA           uint16 = 0x0063
+	TLS_DHE_RSA_WITH_DES_CBC_SHA                      uint16 = 0x0015
+	TLS_DHE_DSS_WITH_DES_CBC_SHA                      uint16 = 0x0012
+	TLS_DH_RSA_WITH_DES_CBC_SHA                       uint16 = 0x000f
+	TLS_DH_DSS_WITH_DES_CBC_SHA                       uint16 = 0x000c
+	TLS_DH_anon_WITH_DES_CBC_SHA                      uint16 = 0x001a
+	TLS_RSA_WITH_DES_CBC_SHA                          uint16 = 0x0009
+	TLS_RSA_EXPORT1024_WITH_DES_CBC_SHA               uint16 = 0x0062
+	TLS_RSA_EXPORT1024_WITH_RC2_CBC_56_MD5            uint16 = 0x0061
+	TLS_DHE_DSS_EXPORT1024_WITH_RC4_56_SHA            uint16 = 0x0065
+	TLS_RSA_EXPORT1024_WITH_RC4_56_SHA                uint16 = 0x0064
+	TLS_RSA_EXPORT1024_WITH_RC4_56_MD5                uint16 = 0x0060
+	TLS_DHE_RSA_EXPORT_WITH_DES40_CBC_SHA             uint16 = 0x0014
+	TLS_DHE_DSS_EXPORT_WITH_DES40_CBC_SHA             uint16 = 0x0011
+	TLS_DH_RSA_EXPORT_WITH_DES40_CBC_SHA              uint16 = 0x000e
+	TLS_DH_DSS_EXPORT_WITH_DES40_CBC_SHA              uint16 = 0x000b
+	TLS_DH_anon_EXPORT_WITH_DES40_CBC_SHA             uint16 = 0x0019
+	TLS_RSA_EXPORT_WITH_DES40_CBC_SHA                 uint16 = 0x0008
+	TLS_RSA_EXPORT_WITH_RC2_CBC_40_MD5                uint16 = 0x0006
+	TLS_DH_anon_EXPORT_WITH_RC4_40_MD5                uint16 = 0x0017
+	TLS_RSA_EXPORT_WITH_RC4_40_MD5                    uint16 = 0x0003
+	TLS_ECDHE_RSA_WITH_NULL_SHA                       uint16 = 0xc010
+	TLS_ECDHE_ECDSA_WITH_NULL_SHA                     uint16 = 0xc006
+	TLS_ECDH_anon_WITH_NULL_SHA                       uint16 = 0xc015
+	TLS_ECDH_RSA_WITH_NULL_SHA                        uint16 = 0xc00b
+	TLS_ECDH_ECDSA_WITH_NULL_SHA                      uint16 = 0xc001
+	TLS_RSA_WITH_NULL_SHA256                          uint16 = 0x003b
+	TLS_RSA_WITH_NULL_SHA                             uint16 = 0x0002
+	TLS_RSA_WITH_NULL_MD5                             uint16 = 0x0001
+
+	// TLS 1.3 cipher suites.
+	TLS_AES_128_GCM_SHA256       uint16 = 0x1301
+	TLS_AES_256_GCM_SHA384       uint16 = 0x1302
+	TLS_CHACHA20_POLY1305_SHA256 uint16 = 0x1303
+	TLS_AES_128_CCM_SHA256       uint16 = 0x1304
+	TLS_AES_128_CCM_8_SHA256     uint16 = 0x1305
+
+	// TLS_FALLBACK_SCSV isn't a standard cipher suite but an indicator
+	// that the client is doing version fallback. See RFC 7507.
+	TLS_FALLBACK_SCSV uint16 = 0x5600
+
+	// Legacy names for the corresponding cipher suites with the correct _SHA256
+	// suffix, retained for backward compatibility.
+	TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305   = TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256
+	TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305 = TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256
+)
+
+type cipherSuite struct {
+	name            string
+	opensslname     string
+	MinProtoVersion uint16
+	keyExhange      string
+	authentication  string
+	encryption      string
+	messageAuthCode string
+}
+
+var cipherSuites = map[uint16]cipherSuite{
+	TLS_AES_256_GCM_SHA384:       {"TLS_AES_256_GCM_SHA384", "TLS_AES_256_GCM_SHA384", VersionTLS13, "any", "any", "AESGCM(256)", "AEAD"},
+	TLS_AES_128_GCM_SHA256:       {"TLS_AES_128_GCM_SHA256", "TLS_AES_128_GCM_SHA256", VersionTLS13, "any", "any", "AESGCM(128)", "AEAD"},
+	TLS_CHACHA20_POLY1305_SHA256: {"TLS_CHACHA20_POLY1305_SHA256", "TLS_CHACHA20_POLY1305_SHA256", VersionTLS13, "any", "any", "ChaCha20(256)", "AEAD"},
+	// TLS_AES_128_CCM_SHA256:                            {"TLS_AES_128_CCM_SHA256", "", VersionTLS13, "any", "any", "AESGCM(128)", "AEAD"},
+	// TLS_AES_128_CCM_8_SHA256:                          {"TLS_AES_128_CCM_8_SHA256", "", VersionTLS13, "any", "any", "AESGCM(128)", "AEAD"},
+	TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256:     {"TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256", "ECDHE-ECDSA-CHACHA20-POLY1305", VersionTLS12, "ECDH", "ECDSA", "ChaCha20(256)", "AEAD"},
+	TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256:       {"TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256", "ECDHE-RSA-CHACHA20-POLY1305", VersionTLS12, "ECDH", "RSA", "ChaCha20(256)", "AEAD"},
+	TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256_OLD: {"TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256_OLD", "ECDHE-ECDSA-CHACHA20-POLY1305-OLD", VersionTLS12, "ECDH", "ECDSA", "ChaCha20(256)", "AEAD"},
+	TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256_OLD:   {"TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256_OLD", "ECDHE-RSA-CHACHA20-POLY1305-OLD", VersionTLS12, "ECDH", "RSA", "ChaCha20(256)", "AEAD"},
+	TLS_DHE_RSA_WITH_CHACHA20_POLY1305_SHA256_OLD:     {"TLS_DHE_RSA_WITH_CHACHA20_POLY1305_SHA256_OLD", "DHE-RSA-CHACHA20-POLY1305-OLD", VersionTLS12, "DH", "RSA", "ChaCha20(256)", "AEAD"},
+	TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384:             {"TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384", "ECDHE-RSA-AES256-GCM-SHA384", VersionTLS12, "ECDH", "RSA", "AESGCM(256)", "AEAD"},
+	TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384:           {"TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384", "ECDHE-ECDSA-AES256-GCM-SHA384", VersionTLS12, "ECDH", "ECDSA", "AESGCM(256)", "AEAD"},
+	TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384:             {"TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384", "ECDHE-RSA-AES256-SHA384", VersionTLS12, "ECDH", "RSA", "AES(256)", "SHA384"},
+	TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384:           {"TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384", "ECDHE-ECDSA-AES256-SHA384", VersionTLS12, "ECDH", "ECDSA", "AES(256)", "SHA384"},
+	TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA:                {"TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA", "ECDHE-RSA-AES256-SHA", VersionSSL30, "ECDH", "RSA", "AES(256)", "SHA1"},
+	TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA:              {"TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA", "ECDHE-ECDSA-AES256-SHA", VersionSSL30, "ECDH", "ECDSA", "AES(256)", "SHA1"},
+	// TLS_SRP_SHA_DSS_WITH_AES_256_CBC_SHA:              {"TLS_SRP_SHA_DSS_WITH_AES_256_CBC_SHA", "SRP-DSS-AES-256-CBC-SHA", VersionSSL30, "SRP", "DSS", "AES(256)", "SHA1"},
+	// TLS_SRP_SHA_RSA_WITH_AES_256_CBC_SHA:              {"TLS_SRP_SHA_RSA_WITH_AES_256_CBC_SHA", "SRP-RSA-AES-256-CBC-SHA", VersionSSL30, "SRP", "RSA", "AES(256)", "SHA1"},
+	// TLS_SRP_SHA_WITH_AES_256_CBC_SHA:                  {"TLS_SRP_SHA_WITH_AES_256_CBC_SHA", "SRP-AES-256-CBC-SHA", VersionSSL30, "SRP", "SRP", "AES(256)", "SHA1"},
+	TLS_DH_DSS_WITH_AES_256_GCM_SHA384:           {"TLS_DH_DSS_WITH_AES_256_GCM_SHA384", "DH-DSS-AES256-GCM-SHA384", VersionTLS12, "DH/DSS", "DH", "AESGCM(256)", "AEAD"},
+	TLS_DHE_DSS_WITH_AES_256_GCM_SHA384:          {"TLS_DHE_DSS_WITH_AES_256_GCM_SHA384", "DHE-DSS-AES256-GCM-SHA384", VersionTLS12, "DH", "DSS", "AESGCM(256)", "AEAD"},
+	TLS_DH_RSA_WITH_AES_256_GCM_SHA384:           {"TLS_DH_RSA_WITH_AES_256_GCM_SHA384", "DH-RSA-AES256-GCM-SHA384", VersionTLS12, "DH/RSA", "DH", "AESGCM(256)", "AEAD"},
+	TLS_DHE_RSA_WITH_AES_256_GCM_SHA384:          {"TLS_DHE_RSA_WITH_AES_256_GCM_SHA384", "DHE-RSA-AES256-GCM-SHA384", VersionTLS12, "DH", "RSA", "AESGCM(256)", "AEAD"},
+	TLS_DHE_RSA_WITH_AES_256_CBC_SHA256:          {"TLS_DHE_RSA_WITH_AES_256_CBC_SHA256", "DHE-RSA-AES256-SHA256", VersionTLS12, "DH", "RSA", "AES(256)", "SHA256"},
+	TLS_DHE_DSS_WITH_AES_256_CBC_SHA256:          {"TLS_DHE_DSS_WITH_AES_256_CBC_SHA256", "DHE-DSS-AES256-SHA256", VersionTLS12, "DH", "DSS", "AES(256)", "SHA256"},
+	TLS_DH_RSA_WITH_AES_256_CBC_SHA256:           {"TLS_DH_RSA_WITH_AES_256_CBC_SHA256", "DH-RSA-AES256-SHA256", VersionTLS12, "DH/RSA", "DH", "AES(256)", "SHA256"},
+	TLS_DH_DSS_WITH_AES_256_CBC_SHA256:           {"TLS_DH_DSS_WITH_AES_256_CBC_SHA256", "DH-DSS-AES256-SHA256", VersionTLS12, "DH/DSS", "DH", "AES(256)", "SHA256"},
+	TLS_DHE_RSA_WITH_AES_256_CBC_SHA:             {"TLS_DHE_RSA_WITH_AES_256_CBC_SHA", "DHE-RSA-AES256-SHA", VersionSSL30, "DH", "RSA", "AES(256)", "SHA1"},
+	TLS_DHE_DSS_WITH_AES_256_CBC_SHA:             {"TLS_DHE_DSS_WITH_AES_256_CBC_SHA", "DHE-DSS-AES256-SHA", VersionSSL30, "DH", "DSS", "AES(256)", "SHA1"},
+	TLS_DH_RSA_WITH_AES_256_CBC_SHA:              {"TLS_DH_RSA_WITH_AES_256_CBC_SHA", "DH-RSA-AES256-SHA", VersionSSL30, "DH/RSA", "DH", "AES(256)", "SHA1"},
+	TLS_DH_DSS_WITH_AES_256_CBC_SHA:              {"TLS_DH_DSS_WITH_AES_256_CBC_SHA", "DH-DSS-AES256-SHA", VersionSSL30, "DH/DSS", "DH", "AES(256)", "SHA1"},
+	TLS_ECDHE_RSA_WITH_CAMELLIA_256_CBC_SHA384:   {"TLS_ECDHE_RSA_WITH_CAMELLIA_256_CBC_SHA384", "ECDHE-RSA-CAMELLIA256-SHA384", VersionTLS12, "ECDH", "RSA", "Camellia(256)", "SHA384"},
+	TLS_ECDHE_ECDSA_WITH_CAMELLIA_256_CBC_SHA384: {"TLS_ECDHE_ECDSA_WITH_CAMELLIA_256_CBC_SHA384", "ECDHE-ECDSA-CAMELLIA256-SHA384", VersionTLS12, "ECDH", "ECDSA", "Camellia(256)", "SHA384"},
+	TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA256:     {"TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA256", "DHE-RSA-CAMELLIA256-SHA256", VersionTLS12, "DH", "RSA", "Camellia(256)", "SHA256"},
+	TLS_DHE_DSS_WITH_CAMELLIA_256_CBC_SHA256:     {"TLS_DHE_DSS_WITH_CAMELLIA_256_CBC_SHA256", "DHE-DSS-CAMELLIA256-SHA256", VersionTLS12, "DH", "DSS", "Camellia(256)", "SHA256"},
+	TLS_DH_RSA_WITH_CAMELLIA_256_CBC_SHA256:      {"TLS_DH_RSA_WITH_CAMELLIA_256_CBC_SHA256", "DH-RSA-CAMELLIA256-SHA256", VersionTLS12, "DH/RSA", "DH", "Camellia(256)", "SHA256"},
+	TLS_DH_DSS_WITH_CAMELLIA_256_CBC_SHA256:      {"TLS_DH_DSS_WITH_CAMELLIA_256_CBC_SHA256", "DH-DSS-CAMELLIA256-SHA256", VersionTLS12, "DH/DSS", "DH", "Camellia(256)", "SHA256"},
+	TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA:        {"TLS_DHE_RSA_WITH_CAMELLIA_256_CBC_SHA", "DHE-RSA-CAMELLIA256-SHA", VersionSSL30, "DH", "RSA", "Camellia(256)", "SHA1"},
+	TLS_DHE_DSS_WITH_CAMELLIA_256_CBC_SHA:        {"TLS_DHE_DSS_WITH_CAMELLIA_256_CBC_SHA", "DHE-DSS-CAMELLIA256-SHA", VersionSSL30, "DH", "DSS", "Camellia(256)", "SHA1"},
+	TLS_DH_RSA_WITH_CAMELLIA_256_CBC_SHA:         {"TLS_DH_RSA_WITH_CAMELLIA_256_CBC_SHA", "DH-RSA-CAMELLIA256-SHA", VersionSSL30, "DH/RSA", "DH", "Camellia(256)", "SHA1"},
+	TLS_DH_DSS_WITH_CAMELLIA_256_CBC_SHA:         {"TLS_DH_DSS_WITH_CAMELLIA_256_CBC_SHA", "DH-DSS-CAMELLIA256-SHA", VersionSSL30, "DH/DSS", "DH", "Camellia(256)", "SHA1"},
+	TLS_ECDH_anon_WITH_AES_256_CBC_SHA:           {"TLS_ECDH_anon_WITH_AES_256_CBC_SHA", "AECDH-AES256-SHA", VersionSSL30, "ECDH", "None", "AES(256)", "SHA1"},
+	TLS_DH_anon_WITH_AES_256_GCM_SHA384:          {"TLS_DH_anon_WITH_AES_256_GCM_SHA384", "ADH-AES256-GCM-SHA384", VersionTLS12, "DH", "None", "AESGCM(256)", "AEAD"},
+	TLS_DH_anon_WITH_AES_256_CBC_SHA256:          {"TLS_DH_anon_WITH_AES_256_CBC_SHA256", "ADH-AES256-SHA256", VersionTLS12, "DH", "None", "AES(256)", "SHA256"},
+	TLS_DH_anon_WITH_AES_256_CBC_SHA:             {"TLS_DH_anon_WITH_AES_256_CBC_SHA", "ADH-AES256-SHA", VersionSSL30, "DH", "None", "AES(256)", "SHA1"},
+	TLS_DH_anon_WITH_CAMELLIA_256_CBC_SHA256:     {"TLS_DH_anon_WITH_CAMELLIA_256_CBC_SHA256", "ADH-CAMELLIA256-SHA256", VersionTLS12, "DH", "None", "Camellia(256)", "SHA256"},
+	TLS_DH_anon_WITH_CAMELLIA_256_CBC_SHA:        {"TLS_DH_anon_WITH_CAMELLIA_256_CBC_SHA", "ADH-CAMELLIA256-SHA", VersionSSL30, "DH", "None", "Camellia(256)", "SHA1"},
+	TLS_ECDH_RSA_WITH_AES_256_GCM_SHA384:         {"TLS_ECDH_RSA_WITH_AES_256_GCM_SHA384", "ECDH-RSA-AES256-GCM-SHA384", VersionTLS12, "ECDH/RSA", "ECDH", "AESGCM(256)", "AEAD"},
+	TLS_ECDH_ECDSA_WITH_AES_256_GCM_SHA384:       {"TLS_ECDH_ECDSA_WITH_AES_256_GCM_SHA384", "ECDH-ECDSA-AES256-GCM-SHA384", VersionTLS12, "ECDH/ECDSA", "ECDH", "AESGCM(256)", "AEAD"},
+	TLS_ECDH_RSA_WITH_AES_256_CBC_SHA384:         {"TLS_ECDH_RSA_WITH_AES_256_CBC_SHA384", "ECDH-RSA-AES256-SHA384", VersionTLS12, "ECDH/RSA", "ECDH", "AES(256)", "SHA384"},
+	TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA384:       {"TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA384", "ECDH-ECDSA-AES256-SHA384", VersionTLS12, "ECDH/ECDSA", "ECDH", "AES(256)", "SHA384"},
+	TLS_ECDH_RSA_WITH_AES_256_CBC_SHA:            {"TLS_ECDH_RSA_WITH_AES_256_CBC_SHA", "ECDH-RSA-AES256-SHA", VersionSSL30, "ECDH/RSA", "ECDH", "AES(256)", "SHA1"},
+	TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA:          {"TLS_ECDH_ECDSA_WITH_AES_256_CBC_SHA", "ECDH-ECDSA-AES256-SHA", VersionSSL30, "ECDH/ECDSA", "ECDH", "AES(256)", "SHA1"},
+	TLS_ECDH_RSA_WITH_CAMELLIA_256_CBC_SHA384:    {"TLS_ECDH_RSA_WITH_CAMELLIA_256_CBC_SHA384", "ECDH-RSA-CAMELLIA256-SHA384", VersionTLS12, "ECDH/RSA", "ECDH", "Camellia(256)", "SHA384"},
+	TLS_ECDH_ECDSA_WITH_CAMELLIA_256_CBC_SHA384:  {"TLS_ECDH_ECDSA_WITH_CAMELLIA_256_CBC_SHA384", "ECDH-ECDSA-CAMELLIA256-SHA384", VersionTLS12, "ECDH/ECDSA", "ECDH", "Camellia(256)", "SHA384"},
+	TLS_RSA_WITH_AES_256_GCM_SHA384:              {"TLS_RSA_WITH_AES_256_GCM_SHA384", "AES256-GCM-SHA384", VersionTLS12, "RSA", "RSA", "AESGCM(256)", "AEAD"},
+	TLS_RSA_WITH_AES_256_CBC_SHA256:              {"TLS_RSA_WITH_AES_256_CBC_SHA256", "AES256-SHA256", VersionTLS12, "RSA", "RSA", "AES(256)", "SHA256"},
+	TLS_RSA_WITH_AES_256_CBC_SHA:                 {"TLS_RSA_WITH_AES_256_CBC_SHA", "AES256-SHA", VersionSSL30, "RSA", "RSA", "AES(256)", "SHA1"},
+	TLS_RSA_WITH_CAMELLIA_256_CBC_SHA256:         {"TLS_RSA_WITH_CAMELLIA_256_CBC_SHA256", "CAMELLIA256-SHA256", VersionTLS12, "RSA", "RSA", "Camellia(256)", "SHA256"},
+	TLS_RSA_WITH_CAMELLIA_256_CBC_SHA:            {"TLS_RSA_WITH_CAMELLIA_256_CBC_SHA", "CAMELLIA256-SHA", VersionSSL30, "RSA", "RSA", "Camellia(256)", "SHA1"},
+	// TLS_RSA_PSK_WITH_AES_256_CBC_SHA:                  {"TLS_RSA_PSK_WITH_AES_256_CBC_SHA", "RSA-PSK-AES256-CBC-SHA", VersionSSL30, "RSAPSK", "RSA", "AES(256)", "SHA1"},
+	// TLS_PSK_WITH_AES_256_CBC_SHA:                      {"TLS_PSK_WITH_AES_256_CBC_SHA", "PSK-AES256-CBC-SHA", VersionSSL30, "PSK", "PSK", "AES(256)", "SHA1"},
+	TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256:   {"TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256", "ECDHE-RSA-AES128-GCM-SHA256", VersionTLS12, "ECDH", "RSA", "AESGCM(128)", "AEAD"},
+	TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256: {"TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256", "ECDHE-ECDSA-AES128-GCM-SHA256", VersionTLS12, "ECDH", "ECDSA", "AESGCM(128)", "AEAD"},
+	TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256:   {"TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256", "ECDHE-RSA-AES128-SHA256", VersionTLS12, "ECDH", "RSA", "AES(128)", "SHA256"},
+	TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256: {"TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256", "ECDHE-ECDSA-AES128-SHA256", VersionTLS12, "ECDH", "ECDSA", "AES(128)", "SHA256"},
+	TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA:      {"TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA", "ECDHE-RSA-AES128-SHA", VersionSSL30, "ECDH", "RSA", "AES(128)", "SHA1"},
+	TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA:    {"TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA", "ECDHE-ECDSA-AES128-SHA", VersionSSL30, "ECDH", "ECDSA", "AES(128)", "SHA1"},
+	// TLS_SRP_SHA_DSS_WITH_AES_128_CBC_SHA:              {"TLS_SRP_SHA_DSS_WITH_AES_128_CBC_SHA", "SRP-DSS-AES-128-CBC-SHA", VersionSSL30, "SRP", "DSS", "AES(128)", "SHA1"},
+	// TLS_SRP_SHA_RSA_WITH_AES_128_CBC_SHA:              {"TLS_SRP_SHA_RSA_WITH_AES_128_CBC_SHA", "SRP-RSA-AES-128-CBC-SHA", VersionSSL30, "SRP", "RSA", "AES(128)", "SHA1"},
+	// TLS_SRP_SHA_WITH_AES_128_CBC_SHA:                  {"TLS_SRP_SHA_WITH_AES_128_CBC_SHA", "SRP-AES-128-CBC-SHA", VersionSSL30, "SRP", "SRP", "AES(128)", "SHA1"},
+	TLS_DH_DSS_WITH_AES_128_GCM_SHA256:           {"TLS_DH_DSS_WITH_AES_128_GCM_SHA256", "DH-DSS-AES128-GCM-SHA256", VersionTLS12, "DH/DSS", "DH", "AESGCM(128)", "AEAD"},
+	TLS_DHE_DSS_WITH_AES_128_GCM_SHA256:          {"TLS_DHE_DSS_WITH_AES_128_GCM_SHA256", "DHE-DSS-AES128-GCM-SHA256", VersionTLS12, "DH", "DSS", "AESGCM(128)", "AEAD"},
+	TLS_DH_RSA_WITH_AES_128_GCM_SHA256:           {"TLS_DH_RSA_WITH_AES_128_GCM_SHA256", "DH-RSA-AES128-GCM-SHA256", VersionTLS12, "DH/RSA", "DH", "AESGCM(128)", "AEAD"},
+	TLS_DHE_RSA_WITH_AES_128_GCM_SHA256:          {"TLS_DHE_RSA_WITH_AES_128_GCM_SHA256", "DHE-RSA-AES128-GCM-SHA256", VersionTLS12, "DH", "RSA", "AESGCM(128)", "AEAD"},
+	TLS_DHE_RSA_WITH_AES_128_CBC_SHA256:          {"TLS_DHE_RSA_WITH_AES_128_CBC_SHA256", "DHE-RSA-AES128-SHA256", VersionTLS12, "DH", "RSA", "AES(128)", "SHA256"},
+	TLS_DHE_DSS_WITH_AES_128_CBC_SHA256:          {"TLS_DHE_DSS_WITH_AES_128_CBC_SHA256", "DHE-DSS-AES128-SHA256", VersionTLS12, "DH", "DSS", "AES(128)", "SHA256"},
+	TLS_DH_RSA_WITH_AES_128_CBC_SHA256:           {"TLS_DH_RSA_WITH_AES_128_CBC_SHA256", "DH-RSA-AES128-SHA256", VersionTLS12, "DH/RSA", "DH", "AES(128)", "SHA256"},
+	TLS_DH_DSS_WITH_AES_128_CBC_SHA256:           {"TLS_DH_DSS_WITH_AES_128_CBC_SHA256", "DH-DSS-AES128-SHA256", VersionTLS12, "DH/DSS", "DH", "AES(128)", "SHA256"},
+	TLS_DHE_RSA_WITH_AES_128_CBC_SHA:             {"TLS_DHE_RSA_WITH_AES_128_CBC_SHA", "DHE-RSA-AES128-SHA", VersionSSL30, "DH", "RSA", "AES(128)", "SHA1"},
+	TLS_DHE_DSS_WITH_AES_128_CBC_SHA:             {"TLS_DHE_DSS_WITH_AES_128_CBC_SHA", "DHE-DSS-AES128-SHA", VersionSSL30, "DH", "DSS", "AES(128)", "SHA1"},
+	TLS_DH_RSA_WITH_AES_128_CBC_SHA:              {"TLS_DH_RSA_WITH_AES_128_CBC_SHA", "DH-RSA-AES128-SHA", VersionSSL30, "DH/RSA", "DH", "AES(128)", "SHA1"},
+	TLS_DH_DSS_WITH_AES_128_CBC_SHA:              {"TLS_DH_DSS_WITH_AES_128_CBC_SHA", "DH-DSS-AES128-SHA", VersionSSL30, "DH/DSS", "DH", "AES(128)", "SHA1"},
+	TLS_ECDHE_RSA_WITH_CAMELLIA_128_CBC_SHA256:   {"TLS_ECDHE_RSA_WITH_CAMELLIA_128_CBC_SHA256", "ECDHE-RSA-CAMELLIA128-SHA256", VersionTLS12, "ECDH", "RSA", "Camellia(128)", "SHA256"},
+	TLS_ECDHE_ECDSA_WITH_CAMELLIA_128_CBC_SHA256: {"TLS_ECDHE_ECDSA_WITH_CAMELLIA_128_CBC_SHA256", "ECDHE-ECDSA-CAMELLIA128-SHA256", VersionTLS12, "ECDH", "ECDSA", "Camellia(128)", "SHA256"},
+	TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA256:     {"TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA256", "DHE-RSA-CAMELLIA128-SHA256", VersionTLS12, "DH", "RSA", "Camellia(128)", "SHA256"},
+	TLS_DHE_DSS_WITH_CAMELLIA_128_CBC_SHA256:     {"TLS_DHE_DSS_WITH_CAMELLIA_128_CBC_SHA256", "DHE-DSS-CAMELLIA128-SHA256", VersionTLS12, "DH", "DSS", "Camellia(128)", "SHA256"},
+	TLS_DH_RSA_WITH_CAMELLIA_128_CBC_SHA256:      {"TLS_DH_RSA_WITH_CAMELLIA_128_CBC_SHA256", "DH-RSA-CAMELLIA128-SHA256", VersionTLS12, "DH/RSA", "DH", "Camellia(128)", "SHA256"},
+	TLS_DH_DSS_WITH_CAMELLIA_128_CBC_SHA256:      {"TLS_DH_DSS_WITH_CAMELLIA_128_CBC_SHA256", "DH-DSS-CAMELLIA128-SHA256", VersionTLS12, "DH/DSS", "DH", "Camellia(128)", "SHA256"},
+	TLS_DHE_RSA_WITH_SEED_CBC_SHA:                {"TLS_DHE_RSA_WITH_SEED_CBC_SHA", "DHE-RSA-SEED-SHA", VersionSSL30, "DH", "RSA", "SEED(128)", "SHA1"},
+	TLS_DHE_DSS_WITH_SEED_CBC_SHA:                {"TLS_DHE_DSS_WITH_SEED_CBC_SHA", "DHE-DSS-SEED-SHA", VersionSSL30, "DH", "DSS", "SEED(128)", "SHA1"},
+	TLS_DH_RSA_WITH_SEED_CBC_SHA:                 {"TLS_DH_RSA_WITH_SEED_CBC_SHA", "DH-RSA-SEED-SHA", VersionSSL30, "DH/RSA", "DH", "SEED(128)", "SHA1"},
+	TLS_DH_DSS_WITH_SEED_CBC_SHA:                 {"TLS_DH_DSS_WITH_SEED_CBC_SHA", "DH-DSS-SEED-SHA", VersionSSL30, "DH/DSS", "DH", "SEED(128)", "SHA1"},
+	TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA:        {"TLS_DHE_RSA_WITH_CAMELLIA_128_CBC_SHA", "DHE-RSA-CAMELLIA128-SHA", VersionSSL30, "DH", "RSA", "Camellia(128)", "SHA1"},
+	TLS_DHE_DSS_WITH_CAMELLIA_128_CBC_SHA:        {"TLS_DHE_DSS_WITH_CAMELLIA_128_CBC_SHA", "DHE-DSS-CAMELLIA128-SHA", VersionSSL30, "DH", "DSS", "Camellia(128)", "SHA1"},
+	TLS_DH_RSA_WITH_CAMELLIA_128_CBC_SHA:         {"TLS_DH_RSA_WITH_CAMELLIA_128_CBC_SHA", "DH-RSA-CAMELLIA128-SHA", VersionSSL30, "DH/RSA", "DH", "Camellia(128)", "SHA1"},
+	TLS_DH_DSS_WITH_CAMELLIA_128_CBC_SHA:         {"TLS_DH_DSS_WITH_CAMELLIA_128_CBC_SHA", "DH-DSS-CAMELLIA128-SHA", VersionSSL30, "DH/DSS", "DH", "Camellia(128)", "SHA1"},
+	TLS_ECDH_anon_WITH_AES_128_CBC_SHA:           {"TLS_ECDH_anon_WITH_AES_128_CBC_SHA", "AECDH-AES128-SHA", VersionSSL30, "ECDH", "None", "AES(128)", "SHA1"},
+	TLS_DH_anon_WITH_AES_128_GCM_SHA256:          {"TLS_DH_anon_WITH_AES_128_GCM_SHA256", "ADH-AES128-GCM-SHA256", VersionTLS12, "DH", "None", "AESGCM(128)", "AEAD"},
+	TLS_DH_anon_WITH_AES_128_CBC_SHA256:          {"TLS_DH_anon_WITH_AES_128_CBC_SHA256", "ADH-AES128-SHA256", VersionTLS12, "DH", "None", "AES(128)", "SHA256"},
+	TLS_DH_anon_WITH_AES_128_CBC_SHA:             {"TLS_DH_anon_WITH_AES_128_CBC_SHA", "ADH-AES128-SHA", VersionSSL30, "DH", "None", "AES(128)", "SHA1"},
+	TLS_DH_anon_WITH_CAMELLIA_128_CBC_SHA256:     {"TLS_DH_anon_WITH_CAMELLIA_128_CBC_SHA256", "ADH-CAMELLIA128-SHA256", VersionTLS12, "DH", "None", "Camellia(128)", "SHA256"},
+	TLS_DH_anon_WITH_SEED_CBC_SHA:                {"TLS_DH_anon_WITH_SEED_CBC_SHA", "ADH-SEED-SHA", VersionSSL30, "DH", "None", "SEED(128)", "SHA1"},
+	TLS_DH_anon_WITH_CAMELLIA_128_CBC_SHA:        {"TLS_DH_anon_WITH_CAMELLIA_128_CBC_SHA", "ADH-CAMELLIA128-SHA", VersionSSL30, "DH", "None", "Camellia(128)", "SHA1"},
+	TLS_ECDH_RSA_WITH_AES_128_GCM_SHA256:         {"TLS_ECDH_RSA_WITH_AES_128_GCM_SHA256", "ECDH-RSA-AES128-GCM-SHA256", VersionTLS12, "ECDH/RSA", "ECDH", "AESGCM(128)", "AEAD"},
+	TLS_ECDH_ECDSA_WITH_AES_128_GCM_SHA256:       {"TLS_ECDH_ECDSA_WITH_AES_128_GCM_SHA256", "ECDH-ECDSA-AES128-GCM-SHA256", VersionTLS12, "ECDH/ECDSA", "ECDH", "AESGCM(128)", "AEAD"},
+	TLS_ECDH_RSA_WITH_AES_128_CBC_SHA256:         {"TLS_ECDH_RSA_WITH_AES_128_CBC_SHA256", "ECDH-RSA-AES128-SHA256", VersionTLS12, "ECDH/RSA", "ECDH", "AES(128)", "SHA256"},
+	TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA256:       {"TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA256", "ECDH-ECDSA-AES128-SHA256", VersionTLS12, "ECDH/ECDSA", "ECDH", "AES(128)", "SHA256"},
+	TLS_ECDH_RSA_WITH_AES_128_CBC_SHA:            {"TLS_ECDH_RSA_WITH_AES_128_CBC_SHA", "ECDH-RSA-AES128-SHA", VersionSSL30, "ECDH/RSA", "ECDH", "AES(128)", "SHA1"},
+	TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA:          {"TLS_ECDH_ECDSA_WITH_AES_128_CBC_SHA", "ECDH-ECDSA-AES128-SHA", VersionSSL30, "ECDH/ECDSA", "ECDH", "AES(128)", "SHA1"},
+	TLS_ECDH_RSA_WITH_CAMELLIA_128_CBC_SHA256:    {"TLS_ECDH_RSA_WITH_CAMELLIA_128_CBC_SHA256", "ECDH-RSA-CAMELLIA128-SHA256", VersionTLS12, "ECDH/RSA", "ECDH", "Camellia(128)", "SHA256"},
+	TLS_ECDH_ECDSA_WITH_CAMELLIA_128_CBC_SHA256:  {"TLS_ECDH_ECDSA_WITH_CAMELLIA_128_CBC_SHA256", "ECDH-ECDSA-CAMELLIA128-SHA256", VersionTLS12, "ECDH/ECDSA", "ECDH", "Camellia(128)", "SHA256"},
+	TLS_RSA_WITH_AES_128_GCM_SHA256:              {"TLS_RSA_WITH_AES_128_GCM_SHA256", "AES128-GCM-SHA256", VersionTLS12, "RSA", "RSA", "AESGCM(128)", "AEAD"},
+	TLS_RSA_WITH_AES_128_CBC_SHA256:              {"TLS_RSA_WITH_AES_128_CBC_SHA256", "AES128-SHA256", VersionTLS12, "RSA", "RSA", "AES(128)", "SHA256"},
+	TLS_RSA_WITH_AES_128_CBC_SHA:                 {"TLS_RSA_WITH_AES_128_CBC_SHA", "AES128-SHA", VersionSSL30, "RSA", "RSA", "AES(128)", "SHA1"},
+	TLS_RSA_WITH_CAMELLIA_128_CBC_SHA256:         {"TLS_RSA_WITH_CAMELLIA_128_CBC_SHA256", "CAMELLIA128-SHA256", VersionTLS12, "RSA", "RSA", "Camellia(128)", "SHA256"},
+	TLS_RSA_WITH_SEED_CBC_SHA:                    {"TLS_RSA_WITH_SEED_CBC_SHA", "SEED-SHA", VersionSSL30, "RSA", "RSA", "SEED(128)", "SHA1"},
+	TLS_RSA_WITH_CAMELLIA_128_CBC_SHA:            {"TLS_RSA_WITH_CAMELLIA_128_CBC_SHA", "CAMELLIA128-SHA", VersionSSL30, "RSA", "RSA", "Camellia(128)", "SHA1"},
+	TLS_RSA_WITH_IDEA_CBC_SHA:                    {"TLS_RSA_WITH_IDEA_CBC_SHA", "IDEA-CBC-SHA", VersionSSL30, "RSA", "RSA", "IDEA(128)", "SHA1"},
+	// TLS_RSA_PSK_WITH_AES_128_CBC_SHA:                  {"TLS_RSA_PSK_WITH_AES_128_CBC_SHA", "RSA-PSK-AES128-CBC-SHA", VersionSSL30, "RSAPSK", "RSA", "AES(128)", "SHA1"},
+	// TLS_PSK_WITH_AES_128_CBC_SHA:                      {"TLS_PSK_WITH_AES_128_CBC_SHA", "PSK-AES128-CBC-SHA", VersionSSL30, "PSK", "PSK", "AES(128)", "SHA1"},
+	TLS_ECDHE_RSA_WITH_RC4_128_SHA:   {"TLS_ECDHE_RSA_WITH_RC4_128_SHA", "ECDHE-RSA-RC4-SHA", VersionSSL30, "ECDH", "RSA", "RC4(128)", "SHA1"},
+	TLS_ECDHE_ECDSA_WITH_RC4_128_SHA: {"TLS_ECDHE_ECDSA_WITH_RC4_128_SHA", "ECDHE-ECDSA-RC4-SHA", VersionSSL30, "ECDH", "ECDSA", "RC4(128)", "SHA1"},
+	TLS_DHE_DSS_WITH_RC4_128_SHA:     {"TLS_DHE_DSS_WITH_RC4_128_SHA", "DHE-DSS-RC4-SHA", VersionSSL30, "DH", "DSS", "RC4(128)", "SHA1"},
+	TLS_ECDH_anon_WITH_RC4_128_SHA:   {"TLS_ECDH_anon_WITH_RC4_128_SHA", "AECDH-RC4-SHA", VersionSSL30, "ECDH", "None", "RC4(128)", "SHA1"},
+	TLS_DH_anon_WITH_RC4_128_MD5:     {"TLS_DH_anon_WITH_RC4_128_MD5", "ADH-RC4-MD5", VersionSSL30, "DH", "None", "RC4(128)", "MD5"},
+	TLS_ECDH_RSA_WITH_RC4_128_SHA:    {"TLS_ECDH_RSA_WITH_RC4_128_SHA", "ECDH-RSA-RC4-SHA", VersionSSL30, "ECDH/RSA", "ECDH", "RC4(128)", "SHA1"},
+	TLS_ECDH_ECDSA_WITH_RC4_128_SHA:  {"TLS_ECDH_ECDSA_WITH_RC4_128_SHA", "ECDH-ECDSA-RC4-SHA", VersionSSL30, "ECDH/ECDSA", "ECDH", "RC4(128)", "SHA1"},
+	TLS_RSA_WITH_RC4_128_SHA:         {"TLS_RSA_WITH_RC4_128_SHA", "RC4-SHA", VersionSSL30, "RSA", "RSA", "RC4(128)", "SHA1"},
+	TLS_RSA_WITH_RC4_128_MD5:         {"TLS_RSA_WITH_RC4_128_MD5", "RC4-MD5", VersionSSL30, "RSA", "RSA", "RC4(128)", "MD5"},
+	// TLS_RSA_PSK_WITH_RC4_128_SHA:                      {"TLS_RSA_PSK_WITH_RC4_128_SHA", "RSA-PSK-RC4-SHA", VersionSSL30, "RSAPSK", "RSA", "RC4(128)", "SHA1"},
+	// TLS_PSK_WITH_RC4_128_SHA:                          {"TLS_PSK_WITH_RC4_128_SHA", "PSK-RC4-SHA", VersionSSL30, "PSK", "PSK", "RC4(128)", "SHA1"},
+	TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA:   {"TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA", "ECDHE-RSA-DES-CBC3-SHA", VersionSSL30, "ECDH", "RSA", "3DES(168)", "SHA1"},
+	TLS_ECDHE_ECDSA_WITH_3DES_EDE_CBC_SHA: {"TLS_ECDHE_ECDSA_WITH_3DES_EDE_CBC_SHA", "ECDHE-ECDSA-DES-CBC3-SHA", VersionSSL30, "ECDH", "ECDSA", "3DES(168)", "SHA1"},
+	// TLS_SRP_SHA_DSS_WITH_3DES_EDE_CBC_SHA:             {"TLS_SRP_SHA_DSS_WITH_3DES_EDE_CBC_SHA", "SRP-DSS-3DES-EDE-CBC-SHA", VersionSSL30, "SRP", "DSS", "3DES(168)", "SHA1"},
+	// TLS_SRP_SHA_RSA_WITH_3DES_EDE_CBC_SHA:             {"TLS_SRP_SHA_RSA_WITH_3DES_EDE_CBC_SHA", "SRP-RSA-3DES-EDE-CBC-SHA", VersionSSL30, "SRP", "RSA", "3DES(168)", "SHA1"},
+	// TLS_SRP_SHA_WITH_3DES_EDE_CBC_SHA:                 {"TLS_SRP_SHA_WITH_3DES_EDE_CBC_SHA", "SRP-3DES-EDE-CBC-SHA", VersionSSL30, "SRP", "SRP", "3DES(168)", "SHA1"},
+	TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA:    {"TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA", "EDH-RSA-DES-CBC3-SHA", VersionSSL30, "DH", "RSA", "3DES(168)", "SHA1"},
+	TLS_DHE_DSS_WITH_3DES_EDE_CBC_SHA:    {"TLS_DHE_DSS_WITH_3DES_EDE_CBC_SHA", "EDH-DSS-DES-CBC3-SHA", VersionSSL30, "DH", "DSS", "3DES(168)", "SHA1"},
+	TLS_DH_RSA_WITH_3DES_EDE_CBC_SHA:     {"TLS_DH_RSA_WITH_3DES_EDE_CBC_SHA", "DH-RSA-DES-CBC3-SHA", VersionSSL30, "DH/RSA", "DH", "3DES(168)", "SHA1"},
+	TLS_DH_DSS_WITH_3DES_EDE_CBC_SHA:     {"TLS_DH_DSS_WITH_3DES_EDE_CBC_SHA", "DH-DSS-DES-CBC3-SHA", VersionSSL30, "DH/DSS", "DH", "3DES(168)", "SHA1"},
+	TLS_ECDH_anon_WITH_3DES_EDE_CBC_SHA:  {"TLS_ECDH_anon_WITH_3DES_EDE_CBC_SHA", "AECDH-DES-CBC3-SHA", VersionSSL30, "ECDH", "None", "3DES(168)", "SHA1"},
+	TLS_DH_anon_WITH_3DES_EDE_CBC_SHA:    {"TLS_DH_anon_WITH_3DES_EDE_CBC_SHA", "ADH-DES-CBC3-SHA", VersionSSL30, "DH", "None", "3DES(168)", "SHA1"},
+	TLS_ECDH_RSA_WITH_3DES_EDE_CBC_SHA:   {"TLS_ECDH_RSA_WITH_3DES_EDE_CBC_SHA", "ECDH-RSA-DES-CBC3-SHA", VersionSSL30, "ECDH/RSA", "ECDH", "3DES(168)", "SHA1"},
+	TLS_ECDH_ECDSA_WITH_3DES_EDE_CBC_SHA: {"TLS_ECDH_ECDSA_WITH_3DES_EDE_CBC_SHA", "ECDH-ECDSA-DES-CBC3-SHA", VersionSSL30, "ECDH/ECDSA", "ECDH", "3DES(168)", "SHA1"},
+	TLS_RSA_WITH_3DES_EDE_CBC_SHA:        {"TLS_RSA_WITH_3DES_EDE_CBC_SHA", "DES-CBC3-SHA", VersionSSL30, "RSA", "RSA", "3DES(168)", "SHA1"},
+	// TLS_RSA_PSK_WITH_3DES_EDE_CBC_SHA:                 {"TLS_RSA_PSK_WITH_3DES_EDE_CBC_SHA", "RSA-PSK-3DES-EDE-CBC-SHA", VersionSSL30, "RSAPSK", "RSA", "3DES(168)", "SHA1"},
+	// TLS_PSK_WITH_3DES_EDE_CBC_SHA:                     {"TLS_PSK_WITH_3DES_EDE_CBC_SHA", "PSK-3DES-EDE-CBC-SHA", VersionSSL30, "PSK", "PSK", "3DES(168)", "SHA1"},
+	TLS_DHE_DSS_EXPORT1024_WITH_DES_CBC_SHA: {"TLS_DHE_DSS_EXPORT1024_WITH_DES_CBC_SHA", "EXP1024-DHE-DSS-DES-CBC-SHA", VersionSSL30, "DH(1024", "DSS", "DES(56)", "SHA1"},
+	TLS_DHE_RSA_WITH_DES_CBC_SHA:            {"TLS_DHE_RSA_WITH_DES_CBC_SHA", "EDH-RSA-DES-CBC-SHA", VersionSSL30, "DH", "RSA", "DES(56)", "SHA1"},
+	TLS_DHE_DSS_WITH_DES_CBC_SHA:            {"TLS_DHE_DSS_WITH_DES_CBC_SHA", "EDH-DSS-DES-CBC-SHA", VersionSSL30, "DH", "DSS", "DES(56)", "SHA1"},
+	TLS_DH_RSA_WITH_DES_CBC_SHA:             {"TLS_DH_RSA_WITH_DES_CBC_SHA", "DH-RSA-DES-CBC-SHA", VersionSSL30, "DH/RSA", "DH", "DES(56)", "SHA1"},
+	TLS_DH_DSS_WITH_DES_CBC_SHA:             {"TLS_DH_DSS_WITH_DES_CBC_SHA", "DH-DSS-DES-CBC-SHA", VersionSSL30, "DH/DSS", "DH", "DES(56)", "SHA1"},
+	TLS_DH_anon_WITH_DES_CBC_SHA:            {"TLS_DH_anon_WITH_DES_CBC_SHA", "ADH-DES-CBC-SHA", VersionSSL30, "DH", "None", "DES(56)", "SHA1"},
+	TLS_RSA_WITH_DES_CBC_SHA:                {"TLS_RSA_WITH_DES_CBC_SHA", "DES-CBC-SHA", VersionSSL30, "RSA", "RSA", "DES(56)", "SHA1"},
+	TLS_RSA_EXPORT1024_WITH_DES_CBC_SHA:     {"TLS_RSA_EXPORT1024_WITH_DES_CBC_SHA", "EXP1024-DES-CBC-SHA", VersionSSL30, "RSA(1024", "RSA", "DES(56)", "SHA1"},
+	TLS_RSA_EXPORT1024_WITH_RC2_CBC_56_MD5:  {"TLS_RSA_EXPORT1024_WITH_RC2_CBC_56_MD5", "EXP1024-RC2-CBC-MD5", VersionSSL30, "RSA(1024", "RSA", "RC2(56)", "MD5 "},
+	TLS_DHE_DSS_EXPORT1024_WITH_RC4_56_SHA:  {"TLS_DHE_DSS_EXPORT1024_WITH_RC4_56_SHA", "EXP1024-DHE-DSS-RC4-SHA", VersionSSL30, "DH(1024", "DSS", "RC4(56)", "SHA1"},
+	TLS_RSA_EXPORT1024_WITH_RC4_56_SHA:      {"TLS_RSA_EXPORT1024_WITH_RC4_56_SHA", "EXP1024-RC4-SHA", VersionSSL30, "RSA(1024", "RSA", "RC4(56)", "SHA1"},
+	TLS_RSA_EXPORT1024_WITH_RC4_56_MD5:      {"TLS_RSA_EXPORT1024_WITH_RC4_56_MD5", "EXP1024-RC4-MD5", VersionSSL30, "RSA(1024", "RSA", "RC4(56)", "MD5 "},
+	TLS_DHE_RSA_EXPORT_WITH_DES40_CBC_SHA:   {"TLS_DHE_RSA_EXPORT_WITH_DES40_CBC_SHA", "EXP-EDH-RSA-DES-CBC-SHA", VersionSSL30, "DH(512", "RSA", "DES(40)", "SHA1"},
+	TLS_DHE_DSS_EXPORT_WITH_DES40_CBC_SHA:   {"TLS_DHE_DSS_EXPORT_WITH_DES40_CBC_SHA", "EXP-EDH-DSS-DES-CBC-SHA", VersionSSL30, "DH(512", "DSS", "DES(40)", "SHA1"},
+	TLS_DH_RSA_EXPORT_WITH_DES40_CBC_SHA:    {"TLS_DH_RSA_EXPORT_WITH_DES40_CBC_SHA", "EXP-DH-RSA-DES-CBC-SHA", VersionSSL30, "DH/RSA", "DH", "DES(40)", "SHA1"},
+	TLS_DH_DSS_EXPORT_WITH_DES40_CBC_SHA:    {"TLS_DH_DSS_EXPORT_WITH_DES40_CBC_SHA", "EXP-DH-DSS-DES-CBC-SHA", VersionSSL30, "DH/DSS", "DH", "DES(40)", "SHA1"},
+	TLS_DH_anon_EXPORT_WITH_DES40_CBC_SHA:   {"TLS_DH_anon_EXPORT_WITH_DES40_CBC_SHA", "EXP-ADH-DES-CBC-SHA", VersionSSL30, "DH(512", "None", "DES(40)", "SHA1"},
+	TLS_RSA_EXPORT_WITH_DES40_CBC_SHA:       {"TLS_RSA_EXPORT_WITH_DES40_CBC_SHA", "EXP-DES-CBC-SHA", VersionSSL30, "RSA(512", "RSA", "DES(40)", "SHA1"},
+	TLS_RSA_EXPORT_WITH_RC2_CBC_40_MD5:      {"TLS_RSA_EXPORT_WITH_RC2_CBC_40_MD5", "EXP-RC2-CBC-MD5", VersionSSL30, "RSA(512", "RSA", "RC2(40)", "MD5 "},
+	TLS_DH_anon_EXPORT_WITH_RC4_40_MD5:      {"TLS_DH_anon_EXPORT_WITH_RC4_40_MD5", "EXP-ADH-RC4-MD5", VersionSSL30, "DH(512", "None", "RC4(40)", "MD5 "},
+	TLS_RSA_EXPORT_WITH_RC4_40_MD5:          {"TLS_RSA_EXPORT_WITH_RC4_40_MD5", "EXP-RC4-MD5", VersionSSL30, "RSA(512", "RSA", "RC4(40)", "MD5 "},
+	TLS_ECDHE_RSA_WITH_NULL_SHA:             {"TLS_ECDHE_RSA_WITH_NULL_SHA", "ECDHE-RSA-NULL-SHA", VersionSSL30, "ECDH", "RSA", "None", "SHA1"},
+	TLS_ECDHE_ECDSA_WITH_NULL_SHA:           {"TLS_ECDHE_ECDSA_WITH_NULL_SHA", "ECDHE-ECDSA-NULL-SHA", VersionSSL30, "ECDH", "ECDSA", "None", "SHA1"},
+	TLS_ECDH_anon_WITH_NULL_SHA:             {"TLS_ECDH_anon_WITH_NULL_SHA", "AECDH-NULL-SHA", VersionSSL30, "ECDH", "None", "None", "SHA1"},
+	TLS_ECDH_RSA_WITH_NULL_SHA:              {"TLS_ECDH_RSA_WITH_NULL_SHA", "ECDH-RSA-NULL-SHA", VersionSSL30, "ECDH/RSA", "ECDH", "None", "SHA1"},
+	TLS_ECDH_ECDSA_WITH_NULL_SHA:            {"TLS_ECDH_ECDSA_WITH_NULL_SHA", "ECDH-ECDSA-NULL-SHA", VersionSSL30, "ECDH/ECDSA", "ECDH", "None", "SHA1"},
+	TLS_RSA_WITH_NULL_SHA256:                {"TLS_RSA_WITH_NULL_SHA256", "NULL-SHA256", VersionTLS12, "RSA", "RSA", "None", "SHA256"},
+	TLS_RSA_WITH_NULL_SHA:                   {"TLS_RSA_WITH_NULL_SHA", "NULL-SHA", VersionSSL30, "RSA", "RSA", "None", "SHA1"},
+	TLS_RSA_WITH_NULL_MD5:                   {"TLS_RSA_WITH_NULL_MD5", "NULL-MD5", VersionSSL30, "RSA", "RSA", "None", "MD5"},
+}
